- move convert_to32 and get_daily_ds_date to xr_utils.py
